requirements.txt
flask
flask_cors
flask-session
flask_bcrypt
flask_sqlalchemy
psycopg2-binary
uuid
PyJWT
authlib
google-cloud-storage
python-dotenv
langchain
langchain_experimental
openai
pinecone-client
flask_cors
PyPDF2
pymongo
bs4
docx
python-docx
PyMuPDF
pytesseract
pandas
openpyxl
moviepy
SpeechRecognition
pydub
gunicorn
redis